+++ Modules/Intent.Modules.Rdbms.Importer/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.13
+
+- Improvement: Better detection of associations to prevent duplicate associations being created.
+
 ### Version 1.0.12
 
 - Improvement: Added `Stored Procedure Operation mapped to Element` option for stored procedure imports. Will explicitly create the mapping from operation to stored procedure.